+++ cases/story_1/run.jl
@@ -180,7 +180,7 @@ using Distributed
 addprocs(2)
 @everywhere using HetaSimulator
 
-mcscn1 = Scenario(model; tspan = (0., 200.), parameters = [:k1=>0.01], saveat = [50., 80., 150.]);
+mcscn1 = Scenario(model, (0., 200.), parameters = [:k1=>0.01], saveat = [50., 80., 150.]);
 mcsim0 = mc(mcscn1, [:k2=>Normal(1e-3,1e-4), :k3=>Normal(1e-4,1e-5)], 20)
 mcsim1 = mc(mcscn1, [:k2=>Normal(1e-3,1e-4), :k3=>Normal(1e-4,1e-5)], 150, parallel_type=EnsembleDistributed())
 =#